+++ src/Components/Core/Presentation/Adaptivity/BreakTimeNotification/BreakTimeNotification.tsx
@@ -83,7 +83,7 @@ export default function BreakTimeNotification({ className }: AdLerUIComponent) {
     <StyledModal
       className={tailwindMerge(className, "")}
       showModal={showModal}
-      onClose={controller.minimizeOrMaximizeBreakNotification}
+      onClose={controller.closeBreakNotification}
       title={translate("pauseTitle").toString()}
     >
       {RenderBreakContent(